refactor(store): delete the file-store-utils re-export barrel
file-store-utils.ts was a single line re-exporting isFileMissingError
and isFileAlreadyExistsError from application/shared/fs-errors.ts, the
real owner. sidecar-file-download-session-store.ts already imported
from the owner directly, so the barrel only made "where do file error
predicates live" have two answers.

CODE_QUALITY_RULES.md and AGENTS.md both named the barrel as the owner,
which pointed contributors at the shell rather than the module.

- Delete the barrel and repoint 3 imports at fs-errors.ts
- Move its test to src/application/shared/fs-errors.test.ts; the
  isFileMissingError({ code: "ENOENT" }) === false case pins that the
  predicate requires a real Error, which is worth keeping
- Correct the owner named in both rule documents
