--- src/adapters/store/file-store-utils.ts
@@ -1,1 +1,0 @@
-export { isFileAlreadyExistsError, isFileMissingError } from "../../application/shared/fs-errors.ts";
